+++ apps/mobile/app/login.tsx
@@ -7,22 +7,12 @@ import * as Crypto from 'expo-crypto';
 import { type Href, useRouter } from 'expo-router';
 import { StatusBar } from 'expo-status-bar';
 import * as WebBrowser from 'expo-web-browser';
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback, useState } from 'react';
 import { Alert, Dimensions, Platform, Pressable, StyleSheet, Text, View } from 'react-native';
-import Animated, {
-  Easing,
-  FadeInDown,
-  FadeInUp,
-  useAnimatedStyle,
-  useSharedValue,
-  withDelay,
-  withRepeat,
-  withSequence,
-  withTiming,
-} from 'react-native-reanimated';
-import Svg, { Circle, Defs, RadialGradient, Stop } from 'react-native-svg';
+import Animated, { FadeInDown, FadeInUp } from 'react-native-reanimated';
 
 import KuguriTitle from '@/assets/icons/kaguri.svg';
+import { AnimatedLoginBackground } from '@/components/AnimatedLoginBackground';
 import { palette } from '@/constants/palette';
 import { fonts } from '@/constants/typography';
 import { useUser } from '@/features/user/UserContext';
@@ -32,7 +22,7 @@ import { getSupabase, isSupabaseConfigured } from '@/lib/supabase';
 
 WebBrowser.maybeCompleteAuthSession();
 
-const { width, height } = Dimensions.get('window');
+const { height } = Dimensions.get('window');
 
 function parseParamsFromUrl(url: string) {
   try {
@@ -66,177 +56,6 @@ function createNonce(length = 32) {
     .slice(0, length)
     .join('');
 }
-
-const AnimatedCircle = Animated.createAnimatedComponent(Circle);
-
-// ふわふわ浮く円のアニメーションコンポーネント（星のエフェクト）
-const FloatingCircle = ({
-  cx,
-  cy,
-  r,
-  duration,
-  delay,
-}: {
-  cx: number;
-  cy: number;
-  r: number;
-  duration: number;
-  delay: number;
-}) => {
-  const translateY = useSharedValue(0);
-  const opacity = useSharedValue(0.3);
-
-  useEffect(() => {
-    translateY.value = withDelay(
-      delay,
-      withSequence(
-        withTiming(-5, { duration: duration / 2, easing: Easing.inOut(Easing.ease) }),
-        withRepeat(withTiming(5, { duration, easing: Easing.inOut(Easing.ease) }), -1, true),
-      ),
-    );
-    opacity.value = withDelay(
-      delay,
-      withRepeat(
-        withTiming(0.9, { duration: duration * 0.8, easing: Easing.inOut(Easing.ease) }),
-        -1,
-        true,
-      ),
-    );
-  }, [delay, duration, opacity, translateY]);
-
-  const animatedProps = useAnimatedStyle(() => ({
-    cy: cy + translateY.value,
-    opacity: opacity.value,
-  }));
-
-  return (
-    <AnimatedCircle
-      cx={cx}
-      cy={animatedProps.cy}
-      r={r}
-      fill='url(#star_grad)'
-      opacity={animatedProps.opacity}
-    />
-  );
-};
-
-// 背景のアニメーション用の定数
-const MESH_COLORS = {
-  base: '#4A5749', // 少し濃いグリーン
-  accent1: '#5B6B5A', // メインのグリーン
-  accent2: '#7A8C79', // 少し明るいグリーン
-  accent3: '#3D473C', // 深いグリーン
-  highlight: '#ffffff15', // 控えめな輝き
-};
-
-const AnimatedMeshBlob = ({
-  initialX,
-  initialY,
-  size,
-  color,
-  duration,
-}: {
-  initialX: number;
-  initialY: number;
-  size: number;
-  color: string;
-  duration: number;
-}) => {
-  const translateX = useSharedValue(0);
-  const translateY = useSharedValue(0);
-  const scale = useSharedValue(1);
-
-  useEffect(() => {
-    translateX.value = withRepeat(
-      withTiming(Math.random() * 100 - 50, { duration, easing: Easing.inOut(Easing.ease) }),
-      -1,
-      true,
-    );
-    translateY.value = withRepeat(
-      withTiming(Math.random() * 100 - 50, {
-        duration: duration * 1.2,
-        easing: Easing.inOut(Easing.ease),
-      }),
-      -1,
-      true,
-    );
-    scale.value = withRepeat(
-      withTiming(1.3, { duration: duration * 0.8, easing: Easing.inOut(Easing.ease) }),
-      -1,
-      true,
-    );
-  }, [duration, scale, translateX, translateY]);
-
-  const animatedStyle = useAnimatedStyle(() => ({
-    transform: [
-      { translateX: translateX.value + initialX },
-      { translateY: translateY.value + initialY },
-      { scale: scale.value },
-    ],
-  }));
-
-  return (
-    <Animated.View
-      style={[
-        styles.blob,
-        {
-          width: size,
-          height: size,
-          borderRadius: size / 2,
-          backgroundColor: color,
-        },
-        animatedStyle,
-      ]}
-    />
-  );
-};
-
-const AnimatedBackground = () => {
-  return (
-    <View style={[StyleSheet.absoluteFill, styles.backgroundContainer]}>
-      {/* メッシュグラデーションを構成する光の塊 */}
-      <AnimatedMeshBlob
-        initialX={-width * 0.2}
-        initialY={-height * 0.1}
-        size={width * 1.2}
-        color={MESH_COLORS.accent1}
-        duration={10000}
-      />
-      <AnimatedMeshBlob
-        initialX={width * 0.5}
-        initialY={height * 0.3}
-        size={width * 1.5}
-        color={MESH_COLORS.accent2}
-        duration={15000}
-      />
-      <AnimatedMeshBlob
-        initialX={width * 0.1}
-        initialY={height * 0.7}
-        size={width * 1.3}
-        color={MESH_COLORS.accent3}
-        duration={12000}
-      />
-
-      {/* 画面全体に薄いオーバーレイをかけて色を馴染ませる */}
-      <View style={[StyleSheet.absoluteFill, styles.overlay]} />
-
-      <Svg height='100%' width='100%' style={StyleSheet.absoluteFill}>
-        <Defs>
-          <RadialGradient id='star_grad' cx='50%' cy='50%' r='50%'>
-            <Stop offset='0%' stopColor='white' stopOpacity='0.8' />
-            <Stop offset='100%' stopColor='white' stopOpacity='0' />
-          </RadialGradient>
-        </Defs>
-        {/* 星のような小さな瞬き */}
-        <FloatingCircle cx={width * 0.3} cy={height * 0.2} r={2} duration={3000} delay={0} />
-        <FloatingCircle cx={width * 0.7} cy={height * 0.4} r={1.5} duration={4000} delay={500} />
-        <FloatingCircle cx={width * 0.1} cy={height * 0.8} r={2.5} duration={5000} delay={1000} />
-        <FloatingCircle cx={width * 0.9} cy={height * 0.1} r={1.2} duration={3500} delay={200} />
-        <FloatingCircle cx={width * 0.5} cy={height * 0.6} r={2} duration={4500} delay={800} />
-      </Svg>
-    </View>
-  );
-};
 
 export default function LoginScreen() {
   const router = useRouter();
@@ -432,7 +251,7 @@ export default function LoginScreen() {
   return (
     <View style={styles.screen}>
       <StatusBar style='light' />
-      <AnimatedBackground />
+      <AnimatedLoginBackground />
 
       <Animated.View entering={FadeInUp.duration(1000).springify()} style={styles.logoContainer}>
         <KuguriTitle
@@ -541,13 +360,6 @@ const styles = StyleSheet.create({
     position: 'relative',
     top: -2,
   },
-  backgroundContainer: {
-    backgroundColor: MESH_COLORS.base,
-    overflow: 'hidden',
-  },
-  blob: {
-    position: 'absolute',
-  },
   button: {
     borderRadius: 36,
     height: 72,
@@ -627,9 +439,6 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     marginBottom: 20,
     marginTop: height * 0.25,
-  },
-  overlay: {
-    backgroundColor: palette.overlay,
   },
   ownerBox: {
     alignItems: 'center',
