+++ apps/mobile/app/owner/login.tsx
@@ -1,9 +1,8 @@
-/* eslint-disable react-native/no-color-literals */
 import { Ionicons } from '@expo/vector-icons';
 import { BlurView } from 'expo-blur';
 import { type Href, useNavigation, useRouter } from 'expo-router';
 import { StatusBar } from 'expo-status-bar';
-import { useEffect, useLayoutEffect, useState } from 'react';
+import { useLayoutEffect, useState } from 'react';
 import {
   Alert,
   Dimensions,
@@ -14,193 +13,18 @@ import {
   TextInput,
   View,
 } from 'react-native';
-import Animated, {
-  Easing,
-  FadeInDown,
-  FadeInUp,
-  useAnimatedStyle,
-  useSharedValue,
-  withDelay,
-  withRepeat,
-  withSequence,
-  withTiming,
-} from 'react-native-reanimated';
-import Svg, { Circle, Defs, RadialGradient, Stop } from 'react-native-svg';
+import Animated, { FadeInDown, FadeInUp } from 'react-native-reanimated';
 
 import KuguriTitle from '@/assets/icons/kaguri.svg';
+import { AnimatedLoginBackground } from '@/components/AnimatedLoginBackground';
 import { palette } from '@/constants/palette';
 import { fonts } from '@/constants/typography';
 import { ensureUserExistsInDB } from '@/lib/auth';
 import { getSupabase, isSupabaseConfigured } from '@/lib/supabase';
 
-const { width, height } = Dimensions.get('window');
+const { height } = Dimensions.get('window');
 
 const isLikelyEmail = (value: string) => /^[^@\s]+@[^@\s]+\.[^@\s]+$/.test(value.trim());
-
-const AnimatedCircle = Animated.createAnimatedComponent(Circle);
-
-const MESH_COLORS = {
-  base: '#4A5749',
-  accent1: '#5B6B5A',
-  accent2: '#7A8C79',
-  accent3: '#3D473C',
-};
-
-const FloatingCircle = ({
-  cx,
-  cy,
-  r,
-  duration,
-  delay,
-}: {
-  cx: number;
-  cy: number;
-  r: number;
-  duration: number;
-  delay: number;
-}) => {
-  const translateY = useSharedValue(0);
-  const opacity = useSharedValue(0.3);
-
-  useEffect(() => {
-    translateY.value = withDelay(
-      delay,
-      withSequence(
-        withTiming(-5, { duration: duration / 2, easing: Easing.inOut(Easing.ease) }),
-        withRepeat(withTiming(5, { duration, easing: Easing.inOut(Easing.ease) }), -1, true),
-      ),
-    );
-    opacity.value = withDelay(
-      delay,
-      withRepeat(
-        withTiming(0.9, { duration: duration * 0.8, easing: Easing.inOut(Easing.ease) }),
-        -1,
-        true,
-      ),
-    );
-  }, [delay, duration, opacity, translateY]);
-
-  const animatedProps = useAnimatedStyle(() => ({
-    cy: cy + translateY.value,
-    opacity: opacity.value,
-  }));
-
-  return (
-    <AnimatedCircle
-      cx={cx}
-      cy={animatedProps.cy}
-      r={r}
-      fill='url(#star_grad)'
-      opacity={animatedProps.opacity}
-    />
-  );
-};
-
-const AnimatedMeshBlob = ({
-  initialX,
-  initialY,
-  size,
-  color,
-  duration,
-}: {
-  initialX: number;
-  initialY: number;
-  size: number;
-  color: string;
-  duration: number;
-}) => {
-  const translateX = useSharedValue(0);
-  const translateY = useSharedValue(0);
-  const scale = useSharedValue(1);
-
-  useEffect(() => {
-    translateX.value = withRepeat(
-      withTiming(Math.random() * 100 - 50, { duration, easing: Easing.inOut(Easing.ease) }),
-      -1,
-      true,
-    );
-    translateY.value = withRepeat(
-      withTiming(Math.random() * 100 - 50, {
-        duration: duration * 1.2,
-        easing: Easing.inOut(Easing.ease),
-      }),
-      -1,
-      true,
-    );
-    scale.value = withRepeat(
-      withTiming(1.3, { duration: duration * 0.8, easing: Easing.inOut(Easing.ease) }),
-      -1,
-      true,
-    );
-  }, [duration, scale, translateX, translateY]);
-
-  const animatedStyle = useAnimatedStyle(() => ({
-    transform: [
-      { translateX: translateX.value + initialX },
-      { translateY: translateY.value + initialY },
-      { scale: scale.value },
-    ],
-  }));
-
-  return (
-    <Animated.View
-      style={[
-        styles.blob,
-        {
-          width: size,
-          height: size,
-          borderRadius: size / 2,
-          backgroundColor: color,
-        },
-        animatedStyle,
-      ]}
-    />
-  );
-};
-
-const AnimatedBackground = () => {
-  return (
-    <View style={[StyleSheet.absoluteFill, styles.backgroundContainer]}>
-      <AnimatedMeshBlob
-        initialX={-width * 0.2}
-        initialY={-height * 0.1}
-        size={width * 1.2}
-        color={MESH_COLORS.accent1}
-        duration={10000}
-      />
-      <AnimatedMeshBlob
-        initialX={width * 0.5}
-        initialY={height * 0.3}
-        size={width * 1.5}
-        color={MESH_COLORS.accent2}
-        duration={15000}
-      />
-      <AnimatedMeshBlob
-        initialX={width * 0.1}
-        initialY={height * 0.7}
-        size={width * 1.3}
-        color={MESH_COLORS.accent3}
-        duration={12000}
-      />
-
-      <View style={[StyleSheet.absoluteFill, styles.overlay]} />
-
-      <Svg height='100%' width='100%' style={StyleSheet.absoluteFill}>
-        <Defs>
-          <RadialGradient id='star_grad' cx='50%' cy='50%' r='50%'>
-            <Stop offset='0%' stopColor='white' stopOpacity='0.8' />
-            <Stop offset='100%' stopColor='white' stopOpacity='0' />
-          </RadialGradient>
-        </Defs>
-        <FloatingCircle cx={width * 0.3} cy={height * 0.2} r={2} duration={3000} delay={0} />
-        <FloatingCircle cx={width * 0.7} cy={height * 0.4} r={1.5} duration={4000} delay={500} />
-        <FloatingCircle cx={width * 0.1} cy={height * 0.8} r={2.5} duration={5000} delay={1000} />
-        <FloatingCircle cx={width * 0.9} cy={height * 0.1} r={1.2} duration={3500} delay={200} />
-        <FloatingCircle cx={width * 0.5} cy={height * 0.6} r={2} duration={4500} delay={800} />
-      </Svg>
-    </View>
-  );
-};
 
 export default function OwnerLoginScreen() {
   const router = useRouter();
@@ -276,7 +100,7 @@ export default function OwnerLoginScreen() {
   return (
     <View style={styles.screen}>
       <StatusBar style='light' />
-      <AnimatedBackground />
+      <AnimatedLoginBackground />
 
       <ScrollView
         contentContainerStyle={styles.scrollContent}
@@ -314,7 +138,7 @@ export default function OwnerLoginScreen() {
                   autoCorrect={false}
                   keyboardType='email-address'
                   placeholder='owner@example.com'
-                  placeholderTextColor='rgba(255, 255, 255, 0.4)'
+                  placeholderTextColor={palette.placeholderText}
                   style={styles.input}
                 />
                 {emailError && <Text style={styles.errorText}>※ 正しく入力してください</Text>}
@@ -327,7 +151,7 @@ export default function OwnerLoginScreen() {
                   onChangeText={setPassword}
                   secureTextEntry
                   placeholder='••••••••'
-                  placeholderTextColor='rgba(255, 255, 255, 0.4)'
+                  placeholderTextColor={palette.placeholderText}
                   style={styles.input}
                 />
               </View>
@@ -385,15 +209,8 @@ const styles = StyleSheet.create({
     fontFamily: fonts.regular,
     fontSize: 14,
   },
-  backgroundContainer: {
-    backgroundColor: MESH_COLORS.base,
-    overflow: 'hidden',
-  },
-  blob: {
-    position: 'absolute',
-  },
   errorText: {
-    color: '#FCA5A5',
+    color: palette.formErrorText,
     fontFamily: fonts.regular,
     fontSize: 12,
     marginTop: 4,
@@ -414,8 +231,8 @@ const styles = StyleSheet.create({
     padding: 24,
   },
   input: {
-    backgroundColor: 'rgba(255, 255, 255, 0.1)',
-    borderColor: 'rgba(255, 255, 255, 0.2)',
+    backgroundColor: palette.inputBg,
+    borderColor: palette.inputBorder,
     borderRadius: 12,
     borderWidth: 1,
     color: palette.white,
@@ -428,15 +245,15 @@ const styles = StyleSheet.create({
     marginBottom: 16,
   },
   label: {
-    color: 'rgba(255, 255, 255, 0.8)',
+    color: palette.labelText,
     fontFamily: fonts.medium,
     fontSize: 13,
     marginBottom: 8,
   },
   loginButton: {
     alignItems: 'center',
-    backgroundColor: 'rgba(255, 255, 255, 0.2)',
-    borderColor: 'rgba(255, 255, 255, 0.3)',
+    backgroundColor: palette.loginButtonBg,
+    borderColor: palette.loginButtonBorder,
     borderRadius: 30,
     borderWidth: 1,
     height: 52,
@@ -447,7 +264,7 @@ const styles = StyleSheet.create({
     opacity: 0.75,
   },
   loginButtonPressed: {
-    backgroundColor: 'rgba(255, 255, 255, 0.25)',
+    backgroundColor: palette.loginButtonPressedBg,
   },
   loginButtonText: {
     color: palette.white,
@@ -461,9 +278,6 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     justifyContent: 'center',
     marginTop: height * 0.12,
-  },
-  overlay: {
-    backgroundColor: palette.overlay,
   },
   screen: {
     flex: 1,
