+++ apps/mobile/constants/palette.ts
@@ -105,6 +105,16 @@ export const palette = {
   glassShadow: 'rgba(0, 0, 0, 0.3)',
   overlay: 'rgba(0, 0, 0, 0.15)',
 
+  // フォーム入力UI（オーナーログインなど）
+  formErrorText: '#FCA5A5',
+  inputBg: 'rgba(255, 255, 255, 0.1)',
+  inputBorder: 'rgba(255, 255, 255, 0.2)',
+  labelText: 'rgba(255, 255, 255, 0.8)',
+  placeholderText: 'rgba(255, 255, 255, 0.4)',
+  loginButtonBg: 'rgba(255, 255, 255, 0.2)',
+  loginButtonBorder: 'rgba(255, 255, 255, 0.3)',
+  loginButtonPressedBg: 'rgba(255, 255, 255, 0.25)',
+
   // 開発者モードUI
   devBoxBg: 'rgba(0, 0, 0, 0.3)',
   devBoxBorder: 'rgba(255, 255, 255, 0.2)',
